# backend/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: OweHost API
  description: |
    OweHost Backend API - A comprehensive hosting panel backend system.
    
    ## Authentication
    
    The API supports two authentication methods:
    
    1. **JWT Bearer Token**: Use the `/api/v1/auth/login` endpoint to obtain tokens
    2. **API Key**: Use the `ApiKey` prefix in the Authorization header
    
    ## Rate Limiting
    
    API requests are rate limited. Check response headers for limit information.
    
  version: 1.0.0
  contact:
    name: OweHost Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /api/v1/users:
    get:
      tags:
        - Users
      summary: List users
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: tenant_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
    post:
      tags:
        - Users
      summary: Create user
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
...
  /api/v1/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '404':
          description: User not found
    patch:
      tags:
        - Users
      summary: Update user
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated
    delete:
      tags:
        - Users
      summary: Delete user
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted
...
  /api/v1/databases:
    get:
      tags:
        - Databases
      summary: List databases
      operationId: listDatabases
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of databases
    post:
      tags:
        - Databases
      summary: Create database
      operationId: createDatabase
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseCreateRequest'
      responses:
        '201':
          description: Database created
  /api/v1/databases/{id}:
    get:
      tags:
        - Databases
      summary: Get database by ID
      operationId: getDatabase
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Database details
    delete:
      tags:
        - Databases
      summary: Delete database
      operationId: deleteDatabase
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Database deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        tenant_id:
          type: string
        username:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [active, suspended, terminated]
        uid:
          type: integer
        gid:
          type: integer
        home_directory:
          type: string
        namespace:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserCreateRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        tenant_id:
          type: string
    UserUpdateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        status:
          type: string
          enum: [active, suspended, terminated]
...
    DatabaseCreateRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        type:
          type: string
          enum: [mysql, postgresql, mariadb]
        charset:
          type: string
        collation:
          type: string
...
security:
  - bearerAuth: []